requirements.txt
selenium
requests
django
djangorestframework
psycopg2

pytest
isort
flake8
black